# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/plaid/webhook": {
      "post": {
        "operationId": "PlaidWebhookController_handle",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Webhook accepted"
          },
          "401": {
            "description": "Signature missing or invalid (production)"
          }
        },
        "summary": "Receive a Plaid webhook event",
        "tags": ["plaid-webhook"]
      }
    },
...
  },
  "info": {
    "title": "Flamingo API",
    "description": "Flamingo REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "admin-jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}